+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
 - Add a library interface to ParallelCluster functionality.
 - Add logging of compute node console output to CloudWatch on compute node bootstrap failure.
 - Add failures field containing failure code and reason to `describe-cluster` output when cluster creation fails.
+- Add support for US isolated regions: us-iso-* and us-isob-*.
 
 **CHANGES**
 - Upgrade Slurm to version 22.05.8.